requirements.txt
faker==20.1.0
pandas==2.1.4
psycopg2-binary==2.9.9
hdfs==2.7.3
pyhive==0.7.0
presto-python-client==0.8.4
python-dotenv==1.0.0